Move leaderboard to own component
